refactor: 도메인 모델 구조 및 enum 접근 지정자 변경
- Badge, DiscountType 도메인 관련 열거형을 model 패키지로 이동
- Orders, OrderItem, MenuItem, ReservationDate model 패키지로 이동
- ControllerErrorMessages,DisCountConfig enum public 접근지정자 삭제
