+++ src/main/java/christmas/Application.java
@@ -7,9 +7,8 @@ import view.inputView.Input;
 import view.outputView.Output;
 
 public class Application {
-    public static void main(String[] args) {
-        // TODO: 프로그램 구현
 
+    public static void main(String[] args) {
         Input input = new ConsoleInput();
         Output output = new ConsoleOutput();
         EventPlannerController eventPlannerController = new EventPlannerController(input, output);
