+++ src/main/java/christmas/domain/discount/ChristmasDayDiscount.java
@@ -4,8 +4,8 @@ import static christmas.PlannerConfig.*;
 import static christmas.domain.discount.DiscountConfig.*;
 
 
-import christmas.domain.Orders;
-import christmas.domain.ReservationDate;
+import christmas.domain.model.Orders;
+import christmas.domain.model.ReservationDate;
 import java.time.LocalDate;
 
 public class ChristmasDayDiscount implements Discount {
