+++ src/main/java/christmas/domain/discount/Discount.java
@@ -1,7 +1,7 @@
 package christmas.domain.discount;
 
-import christmas.domain.Orders;
-import christmas.domain.ReservationDate;
+import christmas.domain.model.Orders;
+import christmas.domain.model.ReservationDate;
 
 public interface Discount {
     int calculateDiscount(Orders orders, ReservationDate date);
