+++ src/main/java/christmas/domain/gift/GiftDto.java
@@ -1,6 +1,6 @@
 package christmas.domain.gift;
 
-import christmas.domain.MenuItem;
+import christmas.domain.model.MenuItem;
 
 public record GiftDto(MenuItem menuItem, int count) {
