+++ src/main/java/christmas/domain/gift/GiftPolicy.java
@@ -1,6 +1,6 @@
 package christmas.domain.gift;
 
-import christmas.domain.Orders;
+import christmas.domain.model.Orders;
 import java.util.List;
 
 public interface GiftPolicy {
