+++ src/test/java/christmas/domain/MenuItemTest.java
@@ -1,5 +1,6 @@
 package christmas.domain;
 
+import christmas.domain.model.MenuItem;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
