+++ src/test/java/christmas/domain/OrderItemTest.java
@@ -1,10 +1,12 @@
 package christmas.domain;
 
+import christmas.domain.model.MenuItem;
+import christmas.domain.model.OrderItem;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.api.DisplayName;
 
-import static christmas.domain.DomainErrorMessages.*;
+import static christmas.domain.model.DomainErrorMessages.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
