+++ src/test/java/christmas/domain/badge/BadgeTest.java
@@ -1,5 +1,6 @@
 package christmas.domain.badge;
 
+import christmas.domain.model.Badge;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.api.DisplayName;
