+++ src/test/java/christmas/domain/discount/WeekendDiscountTest.java
@@ -1,11 +1,13 @@
 package christmas.domain.discount;
 
+import christmas.domain.model.MenuItem;
+import christmas.domain.model.OrderItem;
+import christmas.domain.model.Orders;
+import christmas.domain.model.ReservationDate;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.api.DisplayName;
 import static org.assertj.core.api.Assertions.assertThat;
-
-import christmas.domain.*;
 
 import java.util.Arrays;
 import java.util.stream.Stream;
