+++ src/test/java/christmas/fommater/GiftFormatterTest.java
@@ -1,6 +1,6 @@
 package christmas.fommater;
 
-import christmas.domain.MenuItem;
+import christmas.domain.model.MenuItem;
 import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
